Resolved Severe Bugs
Two bugs resolved. First bugs was in directory changing method "define_template_dir". If users passes a directory name without trailing slash, the slash is not added and thus causes 404 error. Now, it will add the trailing slash regardless of slashes added or not.
Second bug was severe. If a view's name end with 'p', the name is trimmed and 'p' is removed from name. For ex. sign_up.php will become "sign_u". Now, a built in php function "basename" is used to remove extension from the name and the bug is resolved.